wip unified snapping hook
