+++ apps/geoportal/src/app/config/featureFlags.ts
@@ -26,10 +26,6 @@ export const featureFlagConfig: FeatureFlagConfig = {
     default: false,
     alias: "extendedSharing",
   },
-  isSnappingEnabled: {
-    default: true,
-    alias: "snapping",
-  },
   debugTileBoundaries: {
     default: false,
     alias: "debugTileBoundaries",
