--- libraries/commons/measurements/src/lib/Measurements.tsx
@@ -1,432 +1,0 @@
-import React, { useState, useEffect, useContext, useRef, memo } from "react";
-import { TopicMapContext } from "react-cismap/contexts/TopicMapContextProvider";
-import "leaflet/dist/leaflet.css";
-import "leaflet-draw/dist/leaflet.draw.css";
-import L from "leaflet";
-import "leaflet-draw";
-import "leaflet-editable";
-import "./utils/measure";
-import "./utils/measure-path";
-import "leaflet-measure-path/leaflet-measure-path.css";
-import "./styles/m-style.css";
-import useDeviceDetection from "./hooks/useDeviceDetection";
-import { useMapMeasurementsContext } from "./context";
-import { UIModeType, MeasurementShapeDrawing } from "..";
-import { MeasurementsSnapping } from "./components/MeasurementsSnapping";
-import { MeasurementStatusDebug } from "./components/MeasurementStatusDebug";
-
-interface MapMeasurementProps {
-  mode?: UIModeType;
-  polygonActiveIcon?: string;
-  polygonIcon?: string;
-  snappingLayers?: any[]; // MapLibre layers for snapping
-}
-
-export const Measurements = ({
-  mode: propMode,
-  polygonActiveIcon,
-  polygonIcon,
-  snappingLayers,
-}: MapMeasurementProps) => {
-  const { realRoutedMapRef } =
-    useContext<typeof TopicMapContext>(TopicMapContext);
-  const {
-    mode,
-    activeShape,
-    setActiveShape,
-    shapes,
-    setShapes,
-    addShape,
-    deleteAll,
-    setDeleteAll,
-    setUpdateShape,
-    visibleShapes,
-    setVisibleShapes,
-    drawingShape: ifDrawing,
-    setDrawingShape,
-    moveToShape,
-    setMoveToShape,
-    showAll,
-    setShowAll,
-    toggleMeasurementMode: toggleUIMode,
-    setMapMovingEnd,
-    deleteShapeById,
-    updateShapeById,
-    setLastVisibleShapeActive,
-    setDrawingWithLastActiveShape,
-    setActiveShapeIfDrawCancelled,
-    updateAreaOfDrawing,
-    deleteVisibleShapeById,
-    setCurrentDrawHandler,
-    // snappingLatlng - DON'T consume here, causes re-render on every mousemove
-    snappingLatlngRef,
-    subscribeToSnappingLatlng,
-    config,
-    setStatus,
-
-    // looks unuseful
-    setStartDrawing,
-  } = useMapMeasurementsContext();
-
-  // Use context mode if propMode is not provided
-  const currentMode = propMode !== undefined ? propMode : mode;
-
-  const [measureControl, setMeasureControl] = useState<any>(null);
-  const [visiblePolylines, setVisiblePolylines] = useState<(string | number)[]>(
-    []
-  );
-  const [drawingShape, setDrawingLine] = useState(null);
-
-  // Track last valid state for recovery
-  const lastValidStateRef = React.useRef<{
-    activeShape: any;
-    mode: string;
-    wasDrawing: boolean;
-  } | null>(null);
-
-  // Track if map is in valid state
-  const [mapIsValid, setMapIsValid] = React.useState(false);
-
-  const device = useDeviceDetection();
-
-  const toggleMeasurementModeHandler = () => {
-    toggleUIMode();
-  };
-
-  useEffect(() => {
-    const leafletMap = realRoutedMapRef.current?.leafletMap?.leafletElement;
-    if (leafletMap && !measureControl) {
-      const mapExample = leafletMap;
-
-      console.debug(
-        "[Measurements] Initializing measurement control with valid map"
-      );
-
-      const customOptions = {
-        position: "topright",
-        // icon_lineActive: makeMeasureActiveIcon,
-        // icon_lineInactive: makeMeasureIcon,
-        icon_polygonActive: polygonActiveIcon,
-        icon_polygonInactive: polygonIcon,
-        activeShape,
-        mode_btn: `<div id='draw-shape-active' class='measure_button_wrapper'><div class='add_shape'>+</div></div>`,
-        msj_disable_tool: "Do you want to disable the tool?",
-        device,
-        shapes,
-        snappingLatlng: snappingLatlngRef?.current,
-        snappingEnabled: config.snappingEnabled,
-        cbSaveShape: saveShapeHandler,
-        cbUpdateShape: updateShapeHandler,
-        cdDeleteShape: deleteShapeHandler,
-        cbDeleteVisibleShapeById: deleteVisibleShapeByIdHandler,
-        cbVisiblePolylinesChange: visiblePolylinesChange,
-        cbSetDrawingStatus: drawingStatusHandler,
-        cbSetDrawingShape: drawingShapeHandler,
-        measurementOrder: findLargestNumber(shapes),
-        measurementMode: currentMode,
-        cbSetActiveShape: setActiveShapeHandler,
-        cbSetUpdateStatusHandler: setUpdateStatusHandler,
-        cbMapMovingEndHandler: mapMovingEndHandler,
-        cbSaveLastActiveShapeIdBeforeDrawingHandler:
-          saveLastActiveShapeIdBeforeDrawingHandler,
-        cbChangeActiveCanceldShapeId: changeActiveCancelledShapeId,
-        cbToggleMeasurementMode: toggleMeasurementModeHandler,
-        cbUpdateAreaOfDrawingMeasurement: updateAreaOfDrawingMeasurementHandler,
-        cbSetCurrentDrawHandler: setCurrentDrawHandler,
-        cbSetMapStatus: setStatus,
-      };
-
-      const measurePolygonControl = (L.control as any).measurePolygon(
-        customOptions
-      );
-      measurePolygonControl.addTo(mapExample);
-
-      setMeasureControl(measurePolygonControl);
-
-      // Restore previous state if available
-      if (lastValidStateRef.current) {
-        console.debug(
-          "[Measurements] Restoring previous state:",
-          lastValidStateRef.current
-        );
-        const savedState = lastValidStateRef.current;
-
-        // Restore mode if it was in measurement mode
-        if (
-          savedState.mode === "measurement" &&
-          currentMode !== "measurement"
-        ) {
-          // Mode will be restored by parent component
-        }
-
-        // Restore active shape if there was one
-        if (savedState.activeShape && !savedState.wasDrawing) {
-          setTimeout(() => {
-            setActiveShape(savedState.activeShape);
-          }, 100);
-        }
-
-        lastValidStateRef.current = null;
-      }
-    }
-  }, [realRoutedMapRef]);
-
-  // Cleanup on unmount
-  useEffect(() => {
-    return () => {
-      if (measureControl) {
-        console.debug("[Measurements] Cleaning up control on unmount");
-        try {
-          const mapExample =
-            realRoutedMapRef.current?.leafletMap?.leafletElement;
-          if (mapExample) {
-            mapExample.removeControl(measureControl);
-          }
-        } catch (e) {
-          console.warn("[Measurements] Error during cleanup:", e);
-        }
-      }
-    };
-  }, [measureControl, realRoutedMapRef]);
-
-  useEffect(() => {
-    if (measureControl && activeShape) {
-      const shapeCoordinates = shapes.filter((s) => s.shapeId === activeShape);
-      const map = realRoutedMapRef.current?.leafletMap?.leafletElement;
-      if (!map) return;
-
-      if (ifDrawing) {
-        setMoveToShape(null);
-      }
-
-      if (shapeCoordinates[0]?.shapeId && !ifDrawing && !deleteAll) {
-        measureControl.changeColorByActivePolyline(
-          map,
-          shapeCoordinates[0].shapeId
-        );
-      }
-      if (showAll) {
-        const allPolylines = measureControl.getAllPolylines(map);
-        measureControl.fitMapToPolylines(map, allPolylines);
-        setShowAll(false);
-      }
-
-      if (deleteAll) {
-        setMoveToShape(null);
-        measureControl.removePolylineById(map, activeShape);
-        const cleanArr = visibleShapes.filter((m) => m.shapeId !== activeShape);
-        deleteShapeHandler(activeShape);
-        setVisibleShapes(cleanArr);
-
-        const cleanAllArr = shapes.filter((m) => m.shapeId !== activeShape);
-        setShapes(cleanAllArr);
-        setDeleteAll(false);
-        if (measureControl.options.shapes.length === 1) {
-          measureControl.options.shapes = [];
-        }
-        const cleanLocalLefletShapes = measureControl.options.shapes.filter(
-          (m) => m.shapeId !== activeShape
-        );
-
-        measureControl.options.shapes = cleanLocalLefletShapes;
-      }
-      if (moveToShape && !deleteAll) {
-        if (shapeCoordinates.length > 0) {
-          measureControl.showActiveShape(map, shapeCoordinates[0]?.coordinates);
-        }
-      }
-    }
-
-    if (measureControl) {
-      const map = realRoutedMapRef.current?.leafletMap?.leafletElement;
-      if (!map) return;
-      measureControl.changeMeasurementMode(currentMode, map);
-      const shapeCoordinates = shapes.filter((s) => s.shapeId === activeShape);
-      if (shapeCoordinates[0]?.shapeId) {
-        measureControl.changeColorByActivePolyline(
-          map,
-          shapeCoordinates[0].shapeId
-        );
-      }
-
-      if (currentMode === "measurement" && visibleShapes.length === 0) {
-        const visibleShapesIds = measureControl.getVisibleShapeIdsArr(
-          measureControl._map
-        );
-      }
-    }
-  }, [
-    activeShape,
-    measureControl,
-    showAll,
-    deleteAll,
-    ifDrawing,
-    moveToShape,
-    currentMode,
-    realRoutedMapRef,
-  ]);
-
-  // keep snappingLatlng and snappingEnabled in sync with control options
-  useEffect(() => {
-    if (measureControl) {
-      try {
-        console.debug("[Measurements] Syncing snapping options:", {
-          enabled: config.snappingEnabled,
-          hasControl: !!measureControl,
-          snappingLatlng: snappingLatlngRef?.current,
-        });
-
-        // Update snappingEnabled immediately
-        measureControl.options.snappingEnabled = config.snappingEnabled;
-
-        // Initial sync of snappingLatlng
-        measureControl.options.snappingLatlng = config.snappingEnabled
-          ? snappingLatlngRef?.current
-          : null;
-
-        // Subscribe to updates to keep control in sync without re-renders
-        if (subscribeToSnappingLatlng) {
-          const unsubscribe = subscribeToSnappingLatlng((latlng) => {
-            if (measureControl && config.snappingEnabled) {
-              measureControl.options.snappingLatlng = latlng;
-            }
-          });
-          return unsubscribe;
-        }
-      } catch (e) {
-        console.error("[Measurements] Error syncing snapping options:", e);
-      }
-    }
-  }, [
-    measureControl,
-    config.snappingEnabled,
-    snappingLatlngRef,
-    subscribeToSnappingLatlng,
-  ]);
-
-  useEffect(() => {
-    if (measureControl) {
-      const cleanedVisibleArr = filterArrByIds(visiblePolylines, shapes);
-
-      // Preserve drawing shape (5555) if we're in drawing mode
-      const drawingShapeInVisible = visibleShapes.find(
-        (s) => s.shapeId === 5555
-      );
-      if (
-        ifDrawing &&
-        drawingShapeInVisible &&
-        !cleanedVisibleArr.find((s) => s.shapeId === 5555)
-      ) {
-        cleanedVisibleArr.push(drawingShapeInVisible);
-      }
-
-      setVisibleShapes(cleanedVisibleArr);
-      measureControl.changeMeasurementsArr(shapes);
-    }
-  }, [visiblePolylines, shapes, ifDrawing]);
-
-  useEffect(() => {
-    if (drawingShape) {
-      const cleanArr = visibleShapes.filter((m) => m.shapeId !== 5555);
-      setVisibleShapes([...cleanArr, drawingShape]);
-    } else {
-      setLastVisibleShapeActive();
-    }
-  }, [drawingShape]);
-
-  const saveShapeHandler = (layer) => {
-    addShape(layer);
-  };
-  const deleteShapeHandler = (id) => {
-    deleteShapeById(id);
-  };
-  const deleteVisibleShapeByIdHandler = (id) => {
-    deleteVisibleShapeById(id);
-  };
-  const updateShapeHandler = (id, newCoordinates, newDistance, newSquare) => {
-    updateShapeById(id, newCoordinates, newDistance, newSquare);
-  };
-
-  const saveLastActiveShapeIdBeforeDrawingHandler = () => {
-    setDrawingWithLastActiveShape();
-  };
-  const changeActiveCancelledShapeId = () => {
-    setActiveShapeIfDrawCancelled();
-  };
-
-  const visiblePolylinesChange = (arr) => {
-    setVisiblePolylines(arr);
-  };
-
-  const drawingStatusHandler = (status) => {
-    setDrawingShape(status);
-    setStartDrawing(status);
-  };
-
-  const drawingShapeHandler = (draw) => {
-    setDrawingLine(draw);
-  };
-  const setActiveShapeHandler = (id) => {
-    setActiveShape(id);
-    setMoveToShape(null);
-  };
-  const setUpdateStatusHandler = (status) => {
-    setUpdateShape(status);
-  };
-  const mapMovingEndHandler = (status) => {
-    setMapMovingEnd(status);
-  };
-
-  const updateAreaOfDrawingMeasurementHandler = (newArea) => {
-    updateAreaOfDrawing(newArea);
-  };
-
-  // Debug: Log what's causing rerenders
-  React.useEffect(() => {
-    console.debug("[Measurements] Rerender triggered. State:", {
-      activeShape,
-      shapesCount: shapes.length,
-      visibleShapesCount: visibleShapes.length,
-      currentMode,
-      ifDrawing,
-      snappingLatlng: !!snappingLatlngRef?.current,
-    });
-  });
-
-  console.debug("RENDER: [MAPMEASUREMENT] MapMeasurement");
-
-  return (
-    <>
-      <MeasurementStatusDebug />
-      {currentMode === "measurement" && (
-        <MeasurementsSnapping maplibreMaps={snappingLayers || []} />
-      )}
-    </>
-  );
-};
-
-function filterArrByIds(
-  arrIds: (string | number)[],
-  fullArray: MeasurementShapeDrawing[]
-): MeasurementShapeDrawing[] {
-  const finalResult: MeasurementShapeDrawing[] = [];
-  fullArray.forEach((currentItem) => {
-    if (arrIds.includes(currentItem.shapeId)) {
-      finalResult.push(currentItem);
-    }
-  });
-
-  return finalResult;
-}
-
-function findLargestNumber(measurements: MeasurementShapeDrawing[]): number {
-  let largestNumber = 0;
-
-  measurements.forEach((item) => {
-    if (item.number > largestNumber) {
-      largestNumber = item.number;
-    }
-  });
-
-  return largestNumber;
-}
