+++ libraries/commons/measurements/src/lib/utils/measure-path.ts
@@ -172,6 +172,65 @@ export const MeasurePolygon = Control.extend({
       },
     });
 
+    // DIAGNOSTIC: Hook into Leaflet.Draw's internal completion to see what triggers it
+    const originalFinishShape = (this._measureHandler as any)._finishShape;
+    if (originalFinishShape) {
+      (this._measureHandler as any)._finishShape = function (e) {
+        const eventInfo = e
+          ? {
+              type: e.type,
+              originalType: e.originalEvent?.type,
+              pointerType: e.originalEvent?.pointerType,
+              target: e.target?.className,
+              timeStamp: e.timeStamp,
+            }
+          : "no event";
+
+        // CRITICAL: Block Leaflet.Draw's native touch finishing
+        // We handle this manually in vertex-click-handler.ts to ensure consistent behavior
+        if (
+          e &&
+          (e.type?.startsWith("touch") ||
+            e.originalEvent?.type?.startsWith("touch") ||
+            e.originalEvent?.pointerType === "touch")
+        ) {
+          console.warn(
+            "[measure-path] Blocking native _finishShape on touch event",
+            eventInfo
+          );
+          return;
+        }
+
+        return originalFinishShape.apply(this, arguments);
+      };
+    }
+
+    const originalCompleteShape = (this._measureHandler as any).completeShape;
+    if (originalCompleteShape) {
+      (this._measureHandler as any).completeShape = function () {
+        console.warn("[measure-path] 🔴 completeShape() called!", {
+          stack: new Error().stack,
+          vertexCount: this._markers?.length || 0,
+          timestamp: Date.now(),
+        });
+        return originalCompleteShape.apply(this, arguments);
+      };
+    }
+
+    const self = this;
+    const originalAddVertex = (this._measureHandler as any).addVertex;
+    if (originalAddVertex) {
+      (this._measureHandler as any).addVertex = function (latlng) {
+        console.log("[measure-path] addVertex() called", {
+          latlng,
+          currentVertexCount: this._markers?.length || 0,
+          timestamp: Date.now(),
+        });
+        (self as any)._lastVertexAdded = Date.now();
+        return originalAddVertex.apply(this, arguments);
+      };
+    }
+
     this.options.currenLine = this._measureHandler;
     this.options.cbSetCurrentDrawHandler(this._measureHandler);
 
@@ -184,13 +243,44 @@ export const MeasurePolygon = Control.extend({
 
     L.drawLocal.draw.handlers.polyline.tooltip.start =
       "Klicken, um den Startpunkt der Messung zu setzen.";
-    L.drawLocal.draw.handlers.polygon.tooltip.start =
-      "Klicken, um den Startpunkt der Messung zu setzen.";
     L.drawLocal.draw.handlers.polyline.tooltip.cont =
       "Klicken (ggf. mehrmals), um die nächsten Punkte des Linienzuges zu setzen.";
     L.drawLocal.draw.handlers.polyline.tooltip.end = tooltipContent;
 
     this._measureHandler.enable();
+
+    // DIAGNOSTIC: Intercept Leaflet.Draw's built-in dblclick handler
+    const originalDblClick = (this._measureHandler as any)._onMouseDblClick;
+    if (originalDblClick) {
+      console.warn(
+        "[measure-path] Found Leaflet.Draw dblclick handler - intercepting"
+      );
+      (this._measureHandler as any)._onMouseDblClick = function (e) {
+        console.error("[measure-path] LEAFLET.DRAW DBLCLICK HANDLER FIRED", {
+          eventType: e?.type,
+          originalEvent: e?.originalEvent?.type,
+          pointerType: e?.originalEvent?.pointerType,
+          timeStamp: e?.timeStamp,
+          vertexCount: this._markers?.length || 0,
+          timestamp: Date.now(),
+          stack: new Error().stack,
+        });
+        return originalDblClick.apply(this, arguments);
+      };
+    } else {
+      console.log(
+        "[measure-path] No dblclick handler found on _measureHandler"
+      );
+    }
+
+    // DIAGNOSTIC: Log all active event listeners on the map
+    console.log("[measure-path] Active map event listeners:", {
+      hasClick: map.listens("click"),
+      hasDblClick: map.listens("dblclick"),
+      clickCount: map.listens("click", true),
+      dblclickCount: map.listens("dblclick", true),
+      timestamp: Date.now(),
+    });
 
     const latlng =
       this.options.snappingEnabled && this.options.snappingLatlng
@@ -385,8 +475,23 @@ export const MeasurePolygon = Control.extend({
     this._mapClickHandler = (event) => {
       const mode = this.options.measurementMode;
 
+      console.log("[measure-path] Map clicked", {
+        isDrawing: this.options.isDrawing,
+        mode,
+        clickAfterShapeSelection: this.options.clickAfterShapeSelection,
+        isFinishingShape: (this as any)._isFinishingShape,
+        eventType: event.originalEvent?.type,
+        targetClassName: (event.originalEvent?.target as HTMLElement)
+          ?.className,
+        latlng: event.latlng,
+        timestamp: Date.now(),
+      });
+
       // Don't start new measurement if we're finishing one
       if ((this as any)._isFinishingShape) {
+        console.log(
+          "[measure-path] Ignoring map click - currently finishing shape"
+        );
         // Clear flag immediately so next click works
         (this as any)._isFinishingShape = false;
         return;
@@ -406,6 +511,13 @@ export const MeasurePolygon = Control.extend({
     };
 
     this._drawCreatedHandler = (event) => {
+      console.warn("[measure-path] ========== draw:created FIRED ==========", {
+        stack: new Error().stack,
+        layerType: event.layerType,
+        vertexCount: event.layer.getLatLngs?.()?.length || 0,
+        timestamp: Date.now(),
+      });
+
       this.options.isDrawing = false;
       this.options.isDrawingEmpty = true;
 
@@ -489,6 +601,12 @@ export const MeasurePolygon = Control.extend({
         },
         () => (this as any)._lastVertexAdded || 0
       );
+      // Handler is now attached to individual markers in _drawDrawvertexHandler
+      // to ensure we can stop propagation before it reaches the map
+      // map.on("click", this._vertexClickHandler);
+      console.log(
+        "[measure-path] Created vertex handler (will attach to markers)"
+      );
     }
 
     this._drawDrawvertexHandler = (event) => {
@@ -554,6 +672,7 @@ export const MeasurePolygon = Control.extend({
 
             // We don't want to remove 'touchstart' listener we just added, so only remove touchend
             layer.off("touchend");
+            // layer.off("touchstart"); // Don't remove this, we just added our own!
 
             console.log("[measure-path] Removed Leaflet.Draw marker listeners");
           }
@@ -664,11 +783,11 @@ export const MeasurePolygon = Control.extend({
 
     this._mousemoveHandler = (event) => {
       const target = event.originalEvent.target;
-      // const isDesktop = this.options.device === "Desktop" ? true : false;
+      const isDesktop = this.options.device === "Desktop" ? true : false;
       const mode = this.options.measurementMode;
       // this._propagateEventToUnderlyingLayers(map, event, "mouseover");
 
-      if (true) {
+      if (isDesktop) {
         if (!this.options.customTooltip && mode === "measurement") {
           const popupPane = map._panes.popupPane;
 
@@ -744,6 +863,11 @@ export const MeasurePolygon = Control.extend({
     if (this._mousemoveHandler) map.off("mousemove", this._mousemoveHandler);
     if (this._mouseoutHandler) map.off("mouseout", this._mouseoutHandler);
 
+    // Vertex click handler is attached to markers, which are removed automatically
+    // if (this._vertexClickHandler) {
+    //   map.off("click", this._vertexClickHandler);
+    // }
+
     // Remove measure layers
     if (this._measureLayers) {
       this._measureLayers.clearLayers();
@@ -775,7 +899,6 @@ export const MeasurePolygon = Control.extend({
       this.options.isDrawing = false;
     } else {
       this._measureHandler.enable();
-      this.options.isDrawingEmpty = true;
     }
   },
