+++ libraries/commons/measurements/src/lib/utils/vertex-click-handler.ts
@@ -1,13 +1,5 @@
-import type { LeafletMouseEvent, LeafletMap } from "@carma/leaflet";
-import type { Control, Marker } from "leaflet";
-
-const MEASUREMENT_VERTEX_ADD_DELAY_MS = 50;
-
-type MeasureHandler = Control.DrawHandler & {
-  _markers?: Marker[];
-  _finishShape?: () => void;
-  _poly?: { _latlngs?: unknown[] };
-};
+import type { LeafletMouseEvent } from "@carma/leaflet";
+import type { Layer, Control, Marker } from "leaflet";
 
 /**
  * Handler for vertex clicks during measurement drawing
@@ -17,49 +9,21 @@ type MeasureHandler = Control.DrawHandler & {
  * - Middle vertices: No action
  */
 export function createVertexClickHandler(
-  getMeasureHandler: () => MeasureHandler | null | undefined,
+  getMeasureHandler: () =>
+    | (Control.DrawHandler & {
+        _markers?: Marker[];
+        _finishShape?: () => void;
+        _poly?: { _latlngs?: unknown[] };
+      })
+    | null
+    | undefined,
   options: Pick<Control.MeasurePolygonOptions, "shapeMode">,
   getCurrentVertexCount: () => number,
-  map?: LeafletMap,
+  map?: any,
   getIsFinishingShape?: () => boolean,
   setIsFinishingShape?: (value: boolean) => void,
   getLastVertexTimestamp?: () => number
 ) {
-  // Helper to handle common shape finishing logic
-  const finishShape = (
-    measureHandler: MeasureHandler,
-    e: LeafletMouseEvent,
-    debugLabel: string
-  ) => {
-    console.warn(
-      `[measure-path] Finishing ${debugLabel} - triggering finish like double-click`
-    );
-
-    if (setIsFinishingShape) {
-      setIsFinishingShape(true);
-    }
-
-    // Disable the handler after draw:created event fires
-    if (map) {
-      map.once("draw:created", () => {
-        if (measureHandler.disable) {
-          measureHandler.disable();
-          console.log(
-            "[measure-path] Disabled measurement handler after draw:created - ready for new measurement"
-          );
-        }
-        // Reset flag on next tick to allow new measurements immediately
-        // but still block the current bubbling click
-        setTimeout(() => {
-          console.debug(`reset finish shape flag (${debugLabel})`);
-          if (setIsFinishingShape) setIsFinishingShape(false);
-        }, 0);
-      });
-    }
-
-    measureHandler._finishShape?.();
-  };
-
   const handler = function (e: LeafletMouseEvent) {
     const measureHandler = getMeasureHandler();
     if (!measureHandler) {
@@ -72,6 +36,8 @@ export function createVertexClickHandler(
       e.originalEvent.stopPropagation();
       e.originalEvent.preventDefault();
     }
+    // Since we are now attaching this handler directly to the marker,
+    // we can get the marker from e.target instead of searching for it.
     const clickedMarker = e.target as any;
     const clickedHandle = clickedMarker.customHandle;
 
@@ -102,7 +68,7 @@ export function createVertexClickHandler(
     if (isLast && getLastVertexTimestamp) {
       const lastAdded = getLastVertexTimestamp();
       const now = Date.now();
-      if (now - lastAdded < MEASUREMENT_VERTEX_ADD_DELAY_MS) {
+      if (now - lastAdded < 200) {
         console.warn(
           "[measure-path] Ignoring click on last vertex - too soon after addVertex",
           { diff: now - lastAdded }
@@ -123,15 +89,66 @@ export function createVertexClickHandler(
         );
         return;
       }
-
+      console.warn(
+        "[measure-path] Closing polygon - triggering finish like double-click"
+      );
       options.shapeMode = "polygon";
-      finishShape(measureHandler, e, "polygon");
+
+      // Set flag to prevent map click handler from starting new measurement
+      if (setIsFinishingShape) {
+        setIsFinishingShape(true);
+      }
+
+      // Trigger finish like double-click does
+      measureHandler._finishShape?.();
+
+      // Stop propagation to prevent map click handler from starting new measurement
+      e.originalEvent?.stopPropagation?.();
+      e.originalEvent?.preventDefault?.();
+
+      // Disable the handler after draw:created event fires
+      if (map) {
+        map.once("draw:created", () => {
+          if (measureHandler.disable) {
+            measureHandler.disable();
+            console.log(
+              "[measure-path] Disabled measurement handler after draw:created - ready for new measurement"
+            );
+          }
+        });
+      }
       return;
     }
 
     // Last vertex: finish line
     if (isLast) {
-      finishShape(measureHandler, e, "line");
+      console.warn(
+        "[measure-path] Finishing line - will disable handler after completion"
+      );
+
+      // Set flag to prevent map click handler from starting new measurement
+      if (setIsFinishingShape) {
+        setIsFinishingShape(true);
+      }
+
+      measureHandler._finishShape?.();
+
+      // Stop propagation to prevent map click handler from starting new measurement
+      e.originalEvent?.stopPropagation?.();
+      e.originalEvent?.preventDefault?.();
+
+      // Disable the handler after draw:created event fires
+      if (map) {
+        map.once("draw:created", () => {
+          if (measureHandler.disable) {
+            measureHandler.disable();
+            console.log(
+              "[measure-path] Disabled measurement handler after draw:created - ready for new measurement"
+            );
+          }
+          setIsFinishingShape(false);
+        });
+      }
       return;
     }
